Add sent folder

+++ apps/mail/cmps/MailList.jsx
@@ -1,8 +1,20 @@
 import { MailPreview } from "./MailPreview.jsx";
 import { emailService } from "../services/mail.service.js";
 
-export function MailList({ emails, setEmails }) {
+export function MailList({ emails, setEmails, setIsComposeShown }) {
+  function removeForEternity(email) {
+    emailService.remove(email.id).then(() => {
+      setEmails((prevEmails) => {
+        return prevEmails.filter((preEmail) => preEmail.id !== email.id);
+      });
+    });
+  }
+
   function onRemoveMail(email) {
+    if (email.removedAt !== null) {
+      removeForEternity(email);
+      return;
+    }
     emailService.moveToTrash(email).then(() => {
       setEmails((prevEmails) => {
         return prevEmails.filter((preEmail) => preEmail.id !== email.id);
@@ -18,6 +30,7 @@ export function MailList({ emails, setEmails }) {
             email={email}
             key={email.id}
             onRemoveMail={onRemoveMail}
+            setIsComposeShown={setIsComposeShown}
           />
         ))}
       </tbody>
